Refactor DraftPick to DrafPicks context
* Refactor DraftPick to DrafPicks context
* See #797
