+++ lib/ex338/fantasy_leagues/fantasy_league.ex
@@ -15,7 +15,7 @@ defmodule Ex338.FantasyLeagues.FantasyLeague do
     field(:max_flex_spots, :integer)
     belongs_to(:sport_draft, Ex338.FantasyPlayers.SportsLeague)
     has_many(:fantasy_teams, Ex338.FantasyTeams.FantasyTeam)
-    has_many(:draft_picks, Ex338.DraftPick)
+    has_many(:draft_picks, Ex338.DraftPicks.DraftPick)
     has_many(:league_sports, Ex338.FantasyLeagues.LeagueSport)
 
     timestamps()
