+++ lib/ex338/fantasy_players/fantasy_player.ex
@@ -18,7 +18,7 @@ defmodule Ex338.FantasyPlayers.FantasyPlayer do
     belongs_to(:sports_league, Ex338.FantasyPlayers.SportsLeague)
     has_many(:roster_positions, Ex338.RosterPositions.RosterPosition)
     has_many(:fantasy_teams, through: [:roster_positions, :fantasy_team])
-    has_many(:draft_picks, Ex338.DraftPick)
+    has_many(:draft_picks, Ex338.DraftPicks.DraftPick)
     has_many(:in_season_draft_picks, Ex338.InSeasonDraftPick, foreign_key: :drafted_player_id)
     has_many(:waiver_adds, Ex338.Waivers.Waiver, foreign_key: :add_fantasy_player_id)
     has_many(:waivers_drops, Ex338.Waivers.Waiver, foreign_key: :drop_fantasy_player_id)
