+++ lib/ex338_web/live/draft_pick_live.ex
@@ -2,19 +2,19 @@ defmodule Ex338Web.DraftPickLive do
   @moduledoc false
   use Ex338Web, :live_view
 
-  alias Ex338.{DraftPick, FantasyLeagues, Accounts}
+  alias Ex338.{DraftPicks, FantasyLeagues, Accounts}
   alias Ex338Web.DraftPickView
 
   def mount(_params, session, socket) do
     if connected?(socket) do
-      DraftPick.Store.subscribe()
+      DraftPicks.subscribe()
       schedule_refresh()
     end
 
     %{"fantasy_league_id" => fantasy_league_id, "current_user_id" => current_user_id} = session
 
     %{draft_picks: picks, fantasy_teams: teams} =
-      DraftPick.Store.get_picks_for_league(fantasy_league_id)
+      DraftPicks.get_picks_for_league(fantasy_league_id)
 
     filter_params = %{sports_league_id: "", fantasy_team_id: ""}
 
@@ -42,7 +42,7 @@ defmodule Ex338Web.DraftPickLive do
   end
 
   def handle_info(:refresh, socket) do
-    new_data = DraftPick.Store.get_picks_for_league(socket.assigns.fantasy_league.id)
+    new_data = DraftPicks.get_picks_for_league(socket.assigns.fantasy_league.id)
 
     filtered_draft_picks = filter_draft_picks(new_data.draft_picks, socket.assigns)
 
@@ -57,7 +57,7 @@ defmodule Ex338Web.DraftPickLive do
   end
 
   def handle_info({"draft_pick", [:draft_pick | _], _}, socket) do
-    new_data = DraftPick.Store.get_picks_for_league(socket.assigns.fantasy_league.id)
+    new_data = DraftPicks.get_picks_for_league(socket.assigns.fantasy_league.id)
 
     filtered_draft_picks = filter_draft_picks(new_data.draft_picks, socket.assigns)
